chore: rename from sdnotifyextension to sdnotify
